Create a helper for the content list slice
