+++ content/webapp/types/multi-content.ts
@@ -8,6 +8,7 @@ import { Series } from './series';
 import { Guide } from './guides';
 import { Weblink } from './weblinks';
 import { Project } from './projects';
+import { Season } from './seasons';
 
 export type MultiContent =
   | Page
@@ -19,4 +20,5 @@ export type MultiContent =
   | Series
   | Guide
   | Weblink
-  | Project;
+  | Project
+  | Season;
